+++ src/com/marklabs/distributionCoverage/MarginOfferedMetadata.java
@@ -1,86 +1,50 @@
 package com.marklabs.distributionCoverage;
 
-
 public class MarginOfferedMetadata {
 
 	private long id;
 	private String brandName;
-	private double supermarket_mo;
-	private double generalStore_mo;
-	private double kiranaStore_mo;
+	private long supermarket_mo;
+	private long generalStore_mo;
+	private long kiranaStore_mo;
 
-	/**
-	 * @return the id
-	 */
 	public long getId() {
 		return id;
 	}
 
-	/**
-	 * @param id
-	 *            the id to set
-	 */
 	public void setId(long id) {
 		this.id = id;
 	}
 
-	/**
-	 * @return the brandName
-	 */
 	public String getBrandName() {
 		return brandName;
 	}
 
-	/**
-	 * @param brandName
-	 *            the brandName to set
-	 */
 	public void setBrandName(String brandName) {
 		this.brandName = brandName;
 	}
 
-	/**
-	 * @return the supermarket_mo
-	 */
-	public double getSupermarket_mo() {
+	public long getSupermarket_mo() {
 		return supermarket_mo;
 	}
 
-	/**
-	 * @param supermarketMo
-	 *            the supermarket_mo to set
-	 */
-	public void setSupermarket_mo(double supermarketMo) {
+	public void setSupermarket_mo(long supermarketMo) {
 		supermarket_mo = supermarketMo;
 	}
 
-	/**
-	 * @return the generalStore_mo
-	 */
-	public double getGeneralStore_mo() {
+	public long getGeneralStore_mo() {
 		return generalStore_mo;
 	}
 
-	/**
-	 * @param generalStoreMo
-	 *            the generalStore_mo to set
-	 */
-	public void setGeneralStore_mo(double generalStoreMo) {
+	public void setGeneralStore_mo(long generalStoreMo) {
 		generalStore_mo = generalStoreMo;
 	}
 
-	/**
-	 * @return the kiranaStore_mo
-	 */
-	public double getKiranaStore_mo() {
+	public long getKiranaStore_mo() {
 		return kiranaStore_mo;
 	}
 
-	/**
-	 * @param kiranaStoreMo
-	 *            the kiranaStore_mo to set
-	 */
-	public void setKiranaStore_mo(double kiranaStoreMo) {
+	public void setKiranaStore_mo(long kiranaStoreMo) {
 		kiranaStore_mo = kiranaStoreMo;
 	}
